Move helper function and remove unneeded keys
